Services/fleet_gateway/requirements.txt
supabase==2.0.2
gotrue==1.3.0
paho-mqtt==1.6.1
python-dotenv==1.0.1
